+++ lib/active_record_compose/attributes/delegation.rb
@@ -15,22 +15,21 @@ module ActiveRecordCompose
         freeze
       end
 
+      def define_delegated_attribute(klass)
+        klass.delegate(reader, writer, to:, allow_nil:)
+        klass.module_eval <<~RUBY, __FILE__, __LINE__ + 1
+          def #{reader}?
+            query?(#{reader})
+          end
+        RUBY
+      end
+
       # @return [String] The attribute name as string
       def attribute_name = attribute.to_s
 
-      def read_attribute(owner)
-        target = owner.send(to)
-        return nil if target.nil? && allow_nil
-
-        target.public_send(reader)
-      end
-
-      def write_attribute(owner, value)
-        target = owner.send(to)
-        return value if target.nil? && allow_nil
-
-        target.public_send(writer, value)
-        value
+      # @return [Hash<String, Object>]
+      def attribute_hash(model)
+        { attribute_name => model.public_send(attribute) }
       end
 
       private