+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 
 - Omitted Specify instance variables in the `:to` option of `delegate_attribute`.
+- refactor: Integrate attributes defined in `delegate_attribute` into an `AttributeSet`.
 
 ## [0.11.3] - 2025-07-13